Support TCP connections.
With this commit it's possible to connect to the tnfsd via TCP. TCP
connection support multiplexing (so many sessions can be opened in a
single connection). Once disconnected, all related sessions will be
closed.

All functions and structs supporting communication now include a pair
of parameters: client address and client FD. If the FD > 0, it means
it's handled by a TCP connection with given descriptor.